scrapes/requirements.txt
pandas
apache-airflow
psycopg2-binary
pytest
lxml